fix: move Suspect type
